fix(breaking-main): Define a single fail/no-fail job (#43)
Makes sure we have a single job to match against as a github pass
requirement.

When the check job gets skipped we cannot use it as a required check
(since `skipped != successful`), so we need a separate job that succeeds
if `skipped || success`.
I also shortened the job names, since the workflow name already has the
full description.