fix(pr-semver-checks): Checkout merge commit instead of PR commit (#62)
`github.even.pull_request.base_sha` corresponds to the latest commit in
the PR's target, rather than the base where the PR's branch diverged.

When comparing breaking changes we should be using the `merge_commit`
rather than the PR's head. Otherwise, we'd see the any breaking changes
merged to the target in the diff.

Fixes #61 

See
- "Unexpected value for github.event.pull_request.base.sha"
[actions/runner#1689](https://www.github.com/actions/runner/issues/1689)
- Checking out the merge commit is the default behaviour of
`actions/checkout`, we inadvertently undid that
[actions/checkout](https://github.com/actions/checkout?tab=readme-ov-file#checkout-pull-request-head-commit-instead-of-merge-commit)