fix(semver-checks): Use the latest commit in the BASELINE branch (#63)
We were using `base.sha` when fetching the baseline for semver checking,
which corresponds to a commit on which the PR is based, but is not
always the latest one (I haven't seen this specified anywhere, but seen
the behaviour from testing).

This PR changes it to use `base.ref`, so the checkout action always
fetches the latest value.

For the PR's branch we use `merge_commit_sha`, which always corresponds
to the result of merging the diff against the latest value of
`base.ref`.