feat(rs-semver-checks): Ignore failure when a new workspace package gets added (#65)
Semver checks fails if the baseline does not contain all the packages in
the current workspace.

We could do something intelligent like parsing all the existing packages
in the baseline and executing rs-semver-checks for each of them
individually, but that's too much work for a rare edge case.

This change just detects the specific error and posts a better comment
saying that the checks are getting skipped instead.

---------

Co-authored-by: Seyon Sivarajah <seyon.sivarajah@quantinuum.com>