fix(semver-checks): Ensure we use the latest target branch commit for checks (#66)
`github.event.pull_request.base.sha` is weird. It points to "the latest
commit at the target branch, at the moment of the PR creation, or the
last force-push to the PR branch".
https://github.com/orgs/community/discussions/59677

This is clearly not what we want, since we use the post-merge commit as
the HEAD ref.

This PR updates the baseline commit we use to the first parent of that
post-merge commit. This should **always** be the latest commit in the
target branch, so the diff will only include changes from the PR.